Use tokio

+++ src/lib.rs
@@ -1,10 +1,9 @@
 use std::ffi::OsString;
-use std::io::{ErrorKind, Read};
+use std::io::ErrorKind;
 use std::os::unix::ffi::OsStringExt;
-use std::os::unix::process::CommandExt;
-use std::process::{Child, Command, Stdio};
-use std::sync::{Arc, Mutex};
-use std::thread;
+use std::pin::Pin;
+use std::process::Stdio;
+use std::sync::Arc;
 use std::{env::set_current_dir, path::Path};
 
 use anyhow::bail;
@@ -12,6 +11,9 @@ use fs_extra::dir::CopyOptions;
 use nix::sys::wait::{waitpid, WaitPidFlag, WaitStatus};
 use nix::unistd::Pid;
 use tempdir::TempDir;
+use tokio::io::{AsyncRead, AsyncReadExt};
+use tokio::process::{Child, Command};
+use tokio::sync::Mutex;
 
 /// Create a temporary directory filled with a copy of `source_dir`.
 pub fn temp_dir_from_template(source_dir: &Path) -> Result<TempDir, Box<dyn std::error::Error>> {
@@ -34,7 +36,7 @@ pub struct Capture {
 
 impl TemporaryChild {
     /// Spawn child process with optional capturing its output.
-    pub fn spawn(cmd: &mut Command, capture: Capture) -> std::io::Result<Self> {
+    pub async fn spawn(cmd: &mut Command, capture: Capture) -> std::io::Result<Self> {
         if capture.stdout.is_some() {
             cmd.stdout(Stdio::piped());
         }
@@ -56,26 +58,26 @@ impl TemporaryChild {
 
         if let Some(capture_stdout) = capture.stdout {
             let stdout = child.stdout.take().unwrap();
-            spawn_dump_to_string(Box::new(stdout), capture_stdout)
+            spawn_dump_to_string(Box::pin(stdout), capture_stdout).await;
         }
 
         if let Some(capture_stderr) = capture.stderr {
             let stderr = child.stderr.take().unwrap();
-            spawn_dump_to_string(Box::new(stderr), capture_stderr)
+            spawn_dump_to_string(Box::pin(stderr), capture_stderr).await;
         }
 
         Ok(TemporaryChild { child })
     }
 }
 
-fn spawn_dump_to_string(
-    mut stream: Box<dyn std::io::Read + Send + Sync>,
+async fn spawn_dump_to_string(
+    mut stream: Pin<Box<dyn AsyncRead + Send + Sync>>, // Box<dyn std::io::Read + Send + Sync>,
     string: Arc<Mutex<String>>,
 ) {
-    thread::spawn(move || {
+    tokio::spawn(async move {
         let mut buf = [0; 4096];
         loop {
-            let r = stream.read(&mut buf);
+            let r = stream.read(&mut buf).await;
             match r {
                 Err(err) => {
                     if err.kind() == ErrorKind::UnexpectedEof {
@@ -88,7 +90,7 @@ fn spawn_dump_to_string(
                     let s = OsString::from_vec(Vec::from(&buf[..r]));
                     string
                         .lock()
-                        .unwrap()
+                        .await
                         .push_str(s.to_str().expect("Wrong text encoding in child output"));
                 }
             }
@@ -98,29 +100,31 @@ fn spawn_dump_to_string(
 
 impl Drop for TemporaryChild {
     fn drop(&mut self) {
-        // Get the process group ID of the child process
-        let pid = -(self.child.id() as i32); // Negative PID targets process group
+        if let Some(id) = self.child.id() {
+            // Get the process group ID of the child process
+            let pid = -(id as i32); // Negative PID targets process group
 
-        unsafe {
-            libc::kill(pid, libc::SIGTERM);
-        } // Send SIGTERM to the group
+            unsafe {
+                libc::kill(pid, libc::SIGTERM);
+            } // Send SIGTERM to the group
 
-        // Wait for the child process to exit
-        loop {
-            match waitpid(
-                Pid::from_raw(self.child.id() as i32),
-                Some(WaitPidFlag::WNOHANG),
-            ) {
-                Ok(WaitStatus::Exited(_, _)) | Ok(WaitStatus::Signaled(_, _, _)) => break,
-                Ok(_) => continue,
-                Err(_) => break,
+            // Wait for the child process to exit
+            loop {
+                match waitpid(
+                    Pid::from_raw(id as i32),
+                    Some(WaitPidFlag::WNOHANG),
+                ) {
+                    Ok(WaitStatus::Exited(_, _)) | Ok(WaitStatus::Signaled(_, _, _)) => break,
+                    Ok(_) => continue,
+                    Err(_) => break,
+                }
             }
         }
     }
 }
 
-pub fn run_successful_command(cmd: &mut Command) -> anyhow::Result<()> {
-    let status = cmd.status()?;
+pub async fn run_successful_command(cmd: &mut Command) -> anyhow::Result<()> {
+    let status = cmd.status().await?;
     if !status.success() {
         match status.code() {
             Some(code) => bail!("Command failed with exit code: {}.", code),
@@ -130,8 +134,8 @@ pub fn run_successful_command(cmd: &mut Command) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn run_failed_command(cmd: &mut Command) -> anyhow::Result<()> {
-    let status = cmd.status()?;
+pub async fn run_failed_command(cmd: &mut Command) -> anyhow::Result<()> {
+    let status = cmd.status().await?;
     if status.success() {
         bail!("Command succeeded though should have failed.");
     }
